+++ queries/goods_query/tests/test_goods_queries.py
@@ -24,55 +24,6 @@ from test_helpers.clients import DataTestClient
 from users.models import Role, GovUser
 
 
-class GoodsQueryManageStatusTests(DataTestClient):
-    @parameterized.expand([["create_clc_query"], ["create_pv_grading_query"]])
-    def test_set_query_status_to_withdrawn_removes_case_from_queues_users_and_updates_status_success(self, cls_func):
-        """
-        When a case is set to a terminal status, its assigned users, case officer and queues should be removed
-        """
-        query = getattr(self, cls_func)("This is a widget", self.organisation)
-        query.case_officer = self.gov_user
-        query.save()
-        query.queues.set([self.queue])
-        CaseAssignment.objects.create(case=query, queue=self.queue, user=self.gov_user)
-        url = reverse("queries:goods_queries:manage_status", kwargs={"pk": query.pk})
-        data = {"status": "withdrawn"}
-
-        response = self.client.put(url, data, **self.gov_headers)
-        query.refresh_from_db()
-
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(query.status.status, CaseStatusEnum.WITHDRAWN)
-        self.assertEqual(query.queues.count(), 0)
-        self.assertEqual(query.case_officer, None)
-        self.assertEqual(CaseAssignment.objects.filter(case=query).count(), 0)
-
-    def test_case_routing_automation_status_change(self):
-        query = self.create_goods_query("This is a widget", self.organisation, "reason", "reason")
-        query.queues.set([self.queue])
-
-        routing_queue = self.create_queue("new queue", self.team)
-        self.create_routing_rule(
-            self.team.id,
-            routing_queue.id,
-            3,
-            status_id=get_case_status_by_status(CaseStatusEnum.PV).id,
-            additional_rules=[],
-        )
-        self.assertNotEqual(query.status.status, CaseStatusEnum.PV)
-
-        url = reverse("queries:goods_queries:manage_status", kwargs={"pk": query.pk})
-        data = {"status": CaseStatusEnum.PV}
-
-        response = self.client.put(url, data, **self.gov_headers)
-        query.refresh_from_db()
-
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(query.status.status, CaseStatusEnum.PV)
-        self.assertEqual(query.queues.count(), 1)
-        self.assertEqual(query.queues.first().id, routing_queue.id)
-
-
 class ControlListClassificationsQueryCreateTests(DataTestClient):
     def setUp(self):
         super().setUp()
